moved some selection picking code from main.cpp to pixel_perfect_sel.cpp
